+++ backend/package.json
@@ -1,49 +1,25 @@
 {
-  "name": "my-remix-app",
-  "private": true,
-  "sideEffects": false,
-  "type": "module",
+  "name": "@wm-tools/backend",
+  "version": "1.0.0",
   "scripts": {
-    "build": "remix vite:build",
-    "deploy": "wrangler pages deploy ./build/client",
-    "dev": "remix vite:dev --port 5102",
-    "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
-    "start": "wrangler pages dev ./build/client",
-    "typecheck": "tsc",
-    "typegen": "wrangler types"
+    "dev": "wrangler dev",
+    "build": "tsc --project tsconfig.json",
+    "deploy:cloudflare": "wrangler deploy --compatibility-date 2024-09-23 --compatibility-flag nodejs_compat"
   },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": "",
   "dependencies": {
-    "@interledger/open-payments": "^6.13.2",
-    "@paralleldrive/cuid2": "^2.2.2",
-    "@remix-run/cloudflare": "^2.16.5",
-    "@remix-run/cloudflare-pages": "^2.16.5",
-    "@remix-run/react": "^2.16.5",
     "aws4fetch": "^1.0.20",
-    "he": "^1.2.0",
-    "isbot": "^4.1.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.2.0",
-    "sanitize-html": "^2.14.0",
-    "zod": "^3.23.8"
+    "hono": "^4.7.8"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20240512.0",
-    "@remix-run/dev": "^2.16.5",
-    "@types/he": "^1.2.3",
-    "@types/react-dom": "^19.1.2",
-    "@types/sanitize-html": "^2.13.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
-    "eslint": "^8.38.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "typescript": "^5.1.6",
-    "vite": "^6.0.0",
-    "vite-tsconfig-paths": "^4.2.1",
-    "wrangler": "3.57.1"
+    "typescript": "5.1.6",
+    "wrangler": "^4.9.1"
   },
   "engines": {
     "node": ">=20.0.0"
-  }
+  },
+  "packageManager": "pnpm@9.12.1"
 }
