--- backend/public/_routes.json
@@ -1,5 +1,0 @@
-{
-  "version": 1,
-  "include": ["/*"],
-  "exclude": ["/favicon.ico", "/assets/*"]
-}
