+++ backend/src/utils/utils.ts
@@ -1,0 +1,3 @@
+export function walletAddressToKey(walletAddress: string): string {
+    return `${decodeURIComponent(walletAddress).replace('$', '').replace('https://', '')}.json`
+}
