+++ backend/wrangler.toml
@@ -1,0 +1,3 @@
+name = "publisher-tools-api"
+main = "src/index.ts"
+compatibility_date = "2025-05-01"
