+++ frontend/app/routes/create.$type.tsx
@@ -23,7 +23,7 @@ import { validConfigTypes, type ModalType } from '~/lib/presets.js'
 import { tooltips } from '~/lib/tooltips.js'
 import type { ElementConfigType } from '~/lib/types.js'
 import { capitalizeFirstLetter, getDefaultData } from '~/lib/utils.js'
-import { commitSession, getSession } from '../../../backend/app/utils/session.server.js'
+import { commitSession, getSession } from '../utils/session.server.js'
 
 export async function loader({ params, request, context }: LoaderFunctionArgs) {
   const { env } = context.cloudflare
