+++ frontend/package.json
@@ -18,8 +18,10 @@
     "@remix-run/serve": "^2.16.5",
     "@tippyjs/react": "^4.2.6",
     "autoprefixer": "^10.4.20",
+    "aws4fetch": "^1.0.20",
     "axios": "^1.7.7",
     "class-variance-authority": "^0.7.0",
+    "he": "^1.2.0",
     "isbot": "^4.4.0",
     "itty-router": "^5.0.18",
     "postcss": "^8.4.49",
@@ -27,6 +29,7 @@
     "react-click-away-listener": "^2.2.3",
     "react-colorful": "^5.6.1",
     "react-dom": "^18.3.1",
+    "sanitize-html": "^2.14.0",
     "tailwindcss": "^3.4.14",
     "tippy.js": "^6.3.7",
     "uuid": "^9.0.1",
@@ -37,8 +40,10 @@
   "devDependencies": {
     "@remix-run/dev": "^2.14.0",
     "@tailwindcss/forms": "^0.5.9",
+    "@types/he": "^1.2.3",
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.2",
+    "@types/sanitize-html": "^2.15.0",
     "@types/uuid": "^9.0.8",
     "concurrently": "^8.2.2",
     "eslint": "^8.57.1",
